[Myl] Integrate REPL, display lex errors
Hooks up the lexer and parser to the REPL, and currently prints out the
results of those operations. We improve the error handling in the lexer:
instead of asserting, we print the faulty code and mark an error. We
break out of our lex loop and return. The caller should then check for
errors and choose what to do (e.g. don't proceed to parsing).